+++ MaxMind.MinFraud.UnitTest/Request/MinFraudRequestTest.cs
@@ -13,21 +13,21 @@ namespace MaxMind.MinFraud.UnitTest.Request
         public void TestAccount()
         {
             var request = new Transaction(device: Device, account: new Account(userId: "1"));
-            Assert.Equal("1", request.Account.UserId);
+            Assert.Equal("1", request.Account!.UserId);
         }
 
         [Fact]
         public void TestBilling()
         {
             var request = new Transaction(device: Device, billing: new Billing(address: "add"));
-            Assert.Equal("add", request.Billing.Address);
+            Assert.Equal("add", request.Billing!.Address);
         }
 
         [Fact]
         public void TestCreditCard()
         {
             var request = new Transaction(device: Device, creditCard: new CreditCard(bankName: "name"));
-            Assert.Equal("name", request.CreditCard.BankName);
+            Assert.Equal("name", request.CreditCard!.BankName);
         }
 
         [Fact]
@@ -41,35 +41,35 @@ namespace MaxMind.MinFraud.UnitTest.Request
         public void TestEmail()
         {
             var request = new Transaction(device: Device, email: new Email(domain: "test.com"));
-            Assert.Equal("test.com", request.Email.Domain);
+            Assert.Equal("test.com", request.Email!.Domain);
         }
 
         [Fact]
         public void TestEvent()
         {
             var request = new Transaction(device: Device, userEvent: new Event(shopId: "1"));
-            Assert.Equal("1", request.Event.ShopId);
+            Assert.Equal("1", request.Event!.ShopId);
         }
 
         [Fact]
         public void TestOrder()
         {
             var request = new Transaction(device: Device, order: new Order(affiliateId: "af1"));
-            Assert.Equal("af1", request.Order.AffiliateId);
+            Assert.Equal("af1", request.Order!.AffiliateId);
         }
 
         [Fact]
         public void TestPayment()
         {
             var request = new Transaction(device: Device, payment: new Payment(declineCode: "d"));
-            Assert.Equal("d", request.Payment.DeclineCode);
+            Assert.Equal("d", request.Payment!.DeclineCode);
         }
 
         [Fact]
         public void TestShipping()
         {
             var request = new Transaction(device: Device, shipping: new Shipping(lastName: "l"));
-            Assert.Equal("l", request.Shipping.LastName);
+            Assert.Equal("l", request.Shipping!.LastName);
         }
 
         [Fact]
@@ -77,7 +77,7 @@ namespace MaxMind.MinFraud.UnitTest.Request
         {
             var request = new Transaction(device: Device,
                 shoppingCart: new List<ShoppingCartItem> {new ShoppingCartItem(itemId: "1")});
-            Assert.Equal("1", request.ShoppingCart[0].ItemId);
+            Assert.Equal("1", request.ShoppingCart![0].ItemId);
         }
     }
 }