+++ MaxMind.MinFraud/Request/Location.cs
@@ -29,17 +29,17 @@ namespace MaxMind.MinFraud.Request
         /// <param name="phoneNumber">The phone number without the country code for the user’s address.</param>
         /// <param name="phoneCountryCode">The country code for phone number associated with the user’s address.</param>
         protected Location(
-            string firstName = null,
-            string lastName = null,
-            string company = null,
-            string address = null,
-            string address2 = null,
-            string city = null,
-            string region = null,
-            string country = null,
-            string postal = null,
-            string phoneNumber = null,
-            string phoneCountryCode = null
+            string? firstName = null,
+            string? lastName = null,
+            string? company = null,
+            string? address = null,
+            string? address2 = null,
+            string? city = null,
+            string? region = null,
+            string? country = null,
+            string? postal = null,
+            string? phoneNumber = null,
+            string? phoneCountryCode = null
         )
         {
             if (country != null && !CountryRe.IsMatch(country))
@@ -63,69 +63,69 @@ namespace MaxMind.MinFraud.Request
         /// The first name associated with the address.
         /// </summary>
         [JsonProperty("first_name")]
-        public string FirstName { get; protected set; }
+        public string? FirstName { get; protected set; }
 
         /// <summary>
         /// The last name associated with the address.
         /// </summary>
         [JsonProperty("last_name")]
-        public string LastName { get; protected set; }
+        public string? LastName { get; protected set; }
 
         /// <summary>
         /// The company name associated with the address.
         /// </summary>
         [JsonProperty("company")]
-        public string Company { get; protected set; }
+        public string? Company { get; protected set; }
 
         /// <summary>
         /// The first line of the address.
         /// </summary>
         [JsonProperty("address")]
-        public string Address { get; protected set; }
+        public string? Address { get; protected set; }
 
         /// <summary>
         /// The second line of the address.
         /// </summary>
         [JsonProperty("address_2")]
-        public string Address2 { get; protected set; }
+        public string? Address2 { get; protected set; }
 
         /// <summary>
         /// The city associated with the address.
         /// </summary>
         [JsonProperty("city")]
-        public string City { get; protected set; }
+        public string? City { get; protected set; }
 
         /// <summary>
         /// The ISO 3166-2 subdivision code for the region associated
         /// with the address.
         /// </summary>
         [JsonProperty("region")]
-        public string Region { get; protected set; }
+        public string? Region { get; protected set; }
 
         /// <summary>
         /// The ISO 3166-1 alpha-2 country code for the country
         /// associated with the address (e.g., "US")
         /// </summary>
         [JsonProperty("country")]
-        public string Country { get; protected set; }
+        public string? Country { get; protected set; }
 
         /// <summary>
         /// The postal code for associated with the address.
         /// </summary>
         [JsonProperty("postal")]
-        public string Postal { get; protected set; }
+        public string? Postal { get; protected set; }
 
         /// <summary>
         /// The phone country code for the phone number associated with the address.
         /// </summary>
         [JsonProperty("phone_number")]
-        public string PhoneNumber { get; protected set; }
+        public string? PhoneNumber { get; protected set; }
 
         /// <summary>
         /// The phone number, without the country code, associated with the address.
         /// </summary>
         [JsonProperty("phone_country_code")]
-        public string PhoneCountryCode { get; protected set; }
+        public string? PhoneCountryCode { get; protected set; }
 
         /// <summary>
         /// Returns a string that represents the current object.